* 'master' of https://github.com/lgmn2018/swcyapi:
  修改一些功能

+++ src/main/java/com/lgmn/swcyapi/vo/home/HomePageVo.java
@@ -2,6 +2,7 @@ package com.lgmn.swcyapi.vo.home;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -9,8 +10,8 @@ public class HomePageVo {
     private String id;
     private String avatar;
     private String nikeName;
-    private Integer groupPower;
-    private Integer personPower;
+    private BigDecimal groupPower;
+    private BigDecimal personPower;
     private Integer score;
     List<HomeAdVo> homeAdType0Vos;
     List<HomeAdVo> homeAdType1Vos;
